fix Histogram elements
