+++ src/test/java/com/pingcap/tikv/statistics/HistogramTest.java
@@ -1,0 +1,131 @@
+/*
+ *
+ * Copyright 2017 PingCAP, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.pingcap.tikv.statistics;
+
+import org.junit.Test;
+
+/** 
+* TiHistogram Tester.
+* 
+* @author <Authors name> 
+* @since <pre>七月 21, 2017</pre> 
+* @version 1.0 
+*/ 
+public class HistogramTest {
+
+/**
+ * Method: histogramFromStorage(long tableID, long isIndex, long colID, long
+ * numberOfDistinctValue)
+ */
+@Test
+public  void testHistogramFromStorage() throws Exception {
+
+}
+
+/** 
+* 
+* Method: equalRowCount(ByteString values) 
+* 
+*/ 
+@Test
+public void testEqualRowCount() throws Exception { 
+
+} 
+
+/** 
+* 
+* Method: greaterRowCount(ByteString values) 
+* 
+*/ 
+@Test
+public void testGreaterRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: greaterAndEqRowCount(ByteString values) 
+* 
+*/ 
+@Test
+public void testGreaterAndEqRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: lessRowCount(ByteString values) 
+* 
+*/ 
+@Test
+public void testLessRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: lessAndEqRowCount(ByteString values) 
+* 
+*/ 
+@Test
+public void testLessAndEqRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: betweenRowCount(ByteString a, ByteString b) 
+* 
+*/ 
+@Test
+public void testBetweenRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: totalRowCount() 
+* 
+*/ 
+@Test
+public void testTotalRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: bucketRowCount() 
+* 
+*/ 
+@Test
+public void testBucketRowCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+/** 
+* 
+* Method: inBucketBetweenCount() 
+* 
+*/ 
+@Test
+public void testInBucketBetweenCount() throws Exception { 
+//TODO: Test goes here... 
+} 
+
+
+}
